+++ src/main/java/com/sprint/mission/discodeit/FrontEnd/Repository/JCFDiscordRepository.java
@@ -1,11 +1,12 @@
-package com.sprint.mission.discodeit.FrontEnd;
+package com.sprint.mission.discodeit.FrontEnd.Repository;
 
+import com.sprint.mission.discodeit.FrontEnd.DiscordRepository;
 import com.sprint.mission.discodeit.entity.User;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class JCFDiscordRepository implements DiscordRepository{
+public class JCFDiscordRepository implements DiscordRepository {
     private static JCFDiscordRepository instance;
     List<User> list;
