+++ src/main/java/com/sprint/mission/discodeit/FrontEnd/Service/JCFDiscordService.java
@@ -1,9 +1,11 @@
-package com.sprint.mission.discodeit.FrontEnd;
+package com.sprint.mission.discodeit.FrontEnd.Service;
 
 import com.sprint.mission.discodeit.Factory.CreateUserFactory;
+import com.sprint.mission.discodeit.FrontEnd.DiscordRepository;
+import com.sprint.mission.discodeit.FrontEnd.DiscordService;
+import com.sprint.mission.discodeit.FrontEnd.Repository.JCFDiscordRepository;
 import com.sprint.mission.discodeit.entity.User;
 
-import java.util.LinkedList;
 import java.util.List;
 import java.util.Scanner;
