+++ src/main/java/com/sprint/mission/discodeit/Repository/file/FileChannelRepository.java
@@ -1,7 +1,6 @@
 package com.sprint.mission.discodeit.Repository.file;
 
 import com.sprint.mission.discodeit.Repository.ChannelRepository;
-import com.sprint.mission.discodeit.entity.Container.Container;
 import com.sprint.mission.discodeit.entity.Message;
 
 import java.io.*;
