--- src/main/java/com/sprint/mission/discodeit/Repository/jcf/impl/LinkedListJCFChannelRepository.java
@@ -1,17 +1,0 @@
-package com.sprint.mission.discodeit.Repository.jcf.impl;
-
-import com.sprint.mission.discodeit.Repository.jcf.JCFChannelRepository;
-import com.sprint.mission.discodeit.entity.Message;
-
-import java.util.LinkedList;
-
-public class LinkedListJCFChannelRepository extends JCFChannelRepository {
-    public LinkedListJCFChannelRepository() {
-        super.updateMessageList(new LinkedList<>());
-    }
-
-    @Override
-    public void save(Message message) {
-        super.save(message);
-    }
-}
