+++ src/main/java/com/sprint/mission/discodeit/service/ChannelService.java
@@ -2,13 +2,9 @@ package com.sprint.mission.discodeit.service;
 
 import com.sprint.mission.discodeit.entity.Message;
 
-import java.util.List;
 import java.util.UUID;
 
-/** <h3>채널 서비스 </h3><p>
- * 유저는 채널에 메시지를 생성한다. <br>
- * 다른 유저는 메시지를 공유한다.</p>
- */
+
 public interface ChannelService {
     public abstract Message write( UUID channelId);
     public abstract Message write(UUID channelId, String str);
