+++ src/main/java/com/sprint/mission/discodeit/service/UserService.java
@@ -1,6 +1,5 @@
 package com.sprint.mission.discodeit.service;
 
-import com.sprint.mission.discodeit.entity.Message;
 import com.sprint.mission.discodeit.entity.Server;
 
 import java.util.List;
