--- src/main/java/com/sprint/mission/discodeit/service/file/FileMessageService.java
@@ -1,4 +1,0 @@
-package com.sprint.mission.discodeit.service.file;
-
-public class FileMessageService {
-}
