+++ src/main/java/com/sprint/mission/discodeit/service/jcf/JCFServerService.java
@@ -3,7 +3,6 @@ package com.sprint.mission.discodeit.service.jcf;
 import com.sprint.mission.discodeit.Factory.CreateChannalFactory;
 import com.sprint.mission.discodeit.Repository.ServerRepository;
 import com.sprint.mission.discodeit.Repository.jcf.JCFServerRepository;
-import com.sprint.mission.discodeit.Repository.jcf.impl.LinkedListJCFServerRepository;
 import com.sprint.mission.discodeit.entity.Container.Channel;
 import com.sprint.mission.discodeit.entity.Container.Container;
 import com.sprint.mission.discodeit.service.ServerService;
